requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
firebase-admin>=6.2.0
passlib>=1.7.4
python-jose>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.4.2
email-validator>=2.0.0
python-multipart>=0.0.6
google-cloud-storage==3.1.0
requests>=2.28.1
jinja2==3.1.6
bcrypt==3.2.0
pyvips>=2.2.0